+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [X.X.X] - Unreleased
 
+### Added
+- `Proof` model and `ProofType` enum
+- `Proof` property on rows returned by `GetSheet` and `GetReport`
+- `PROOFS` include value for `SheetLevelInclusion` and `ReportInclusion`
+
 ## [7.3.0] - 2026-07-20
 
 ### Added