+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `Proof` model and `ProofType` enum
 - `Proof` property on rows returned by `GetSheet` and `GetReport`
-- `PROOFS` include value for `SheetLevelInclusion` and `ReportInclusion`
+- `PROOFS` include value for `SheetLevelInclusion`, `ReportInclusion`, and `RowInclusion`
 
 ## [7.3.0] - 2026-07-20